Change the BusinessMetadataStore to be interface and add DistributedBusinessMetadataStore and InMemoryBusinessMetadataStore as impls.
