+++ cdap-data-fabric-tests/data/namespaces/default/streams/testCleanupGeneration/config.json
@@ -1,0 +1,1 @@
+{"partitionDuration":3600000,"indexInterval":10000,"ttl":9223372036854775807,"notificationThresholdMB":1024}
