+++ cdap-data-fabric-tests/data/namespaces/default/streams/testCleanupTTL/config.json
@@ -1,0 +1,1 @@
+{"partitionDuration":2000,"indexInterval":10000,"ttl":5000,"notificationThresholdMB":1024}
