+++ cdap-data-fabric-tests/src/test/java/co/cask/cdap/data/stream/DistributedStreamCoordinatorClientTest.java
@@ -15,7 +15,6 @@
  */
 package co.cask.cdap.data.stream;
 
-import co.cask.cdap.common.conf.CConfiguration;
 import co.cask.cdap.common.conf.Constants;
 import co.cask.cdap.common.guice.ConfigModule;
 import co.cask.cdap.common.guice.DiscoveryRuntimeModule;
@@ -27,10 +26,8 @@ import co.cask.cdap.data.runtime.DataSetsModules;
 import co.cask.cdap.data.runtime.TransactionMetricsModule;
 import co.cask.cdap.data.stream.service.InMemoryStreamMetaStore;
 import co.cask.cdap.data.stream.service.StreamMetaStore;
-import co.cask.cdap.data2.datafabric.dataset.service.DatasetService;
 import co.cask.cdap.data2.transaction.stream.StreamAdmin;
 import co.cask.cdap.notifications.feeds.guice.NotificationFeedServiceRuntimeModule;
-import co.cask.tephra.TransactionManager;
 import com.google.inject.AbstractModule;
 import com.google.inject.Guice;
 import com.google.inject.Injector;
@@ -57,16 +54,9 @@ public class DistributedStreamCoordinatorClientTest extends StreamCoordinatorTes
   private static MiniDFSCluster dfsCluster;
   private static StreamAdmin streamAdmin;
   private static StreamCoordinatorClient coordinatorClient;
-  private static TransactionManager txManager;
-  private static DatasetService datasetService;
 
   @BeforeClass
   public static void init() throws IOException {
-    cConf.set(Constants.CFG_LOCAL_DATA_DIR, tmpFolder.newFolder().getAbsolutePath());
-    cConf.set(Constants.AppFabric.OUTPUT_DIR, System.getProperty("java.io.tmpdir"));
-    cConf.set(Constants.AppFabric.TEMP_DIR, System.getProperty("java.io.tmpdir"));
-    cConf.setBoolean(Constants.Dangerous.UNRECOVERABLE_RESET, true);
-
     zkServer = InMemoryZKServer.builder().setDataDir(tmpFolder.newFolder()).build();
     zkServer.startAndWait();
 
@@ -103,11 +93,6 @@ public class DistributedStreamCoordinatorClientTest extends StreamCoordinatorTes
         })
     );
 
-    txManager = injector.getInstance(TransactionManager.class);
-    txManager.startAndWait();
-    datasetService = injector.getInstance(DatasetService.class);
-    datasetService.startAndWait();
-
     zkClient = injector.getInstance(ZKClientService.class);
     zkClient.startAndWait();
 
@@ -119,8 +104,6 @@ public class DistributedStreamCoordinatorClientTest extends StreamCoordinatorTes
 
   @AfterClass
   public static void finish() {
-    datasetService.stopAndWait();
-    txManager.stopAndWait();
     coordinatorClient.stopAndWait();
     dfsCluster.shutdown();
     zkClient.stopAndWait();
