+++ cdap-data-fabric-tests/src/test/java/co/cask/cdap/data/stream/InMemoryStreamCoordinatorClientTest.java
@@ -15,7 +15,6 @@
  */
 package co.cask.cdap.data.stream;
 
-import co.cask.cdap.common.conf.Constants;
 import co.cask.cdap.common.guice.ConfigModule;
 import co.cask.cdap.common.guice.DiscoveryRuntimeModule;
 import co.cask.cdap.common.guice.LocationRuntimeModule;
@@ -26,10 +25,8 @@ import co.cask.cdap.data.runtime.SystemDatasetRuntimeModule;
 import co.cask.cdap.data.runtime.TransactionMetricsModule;
 import co.cask.cdap.data.stream.service.InMemoryStreamMetaStore;
 import co.cask.cdap.data.stream.service.StreamMetaStore;
-import co.cask.cdap.data2.datafabric.dataset.service.DatasetService;
 import co.cask.cdap.data2.transaction.stream.StreamAdmin;
 import co.cask.cdap.notifications.feeds.guice.NotificationFeedServiceRuntimeModule;
-import co.cask.tephra.TransactionManager;
 import com.google.inject.AbstractModule;
 import com.google.inject.Guice;
 import com.google.inject.Injector;
@@ -46,15 +43,9 @@ public class InMemoryStreamCoordinatorClientTest extends StreamCoordinatorTestBa
 
   private static StreamAdmin streamAdmin;
   private static StreamCoordinatorClient coordinatorClient;
-  private static TransactionManager txManager;
-  private static DatasetService datasetService;
 
   @BeforeClass
   public static void init() throws IOException {
-    cConf.set(Constants.CFG_LOCAL_DATA_DIR, tmpFolder.newFolder().getAbsolutePath());
-    cConf.set(Constants.AppFabric.OUTPUT_DIR, System.getProperty("java.io.tmpdir"));
-    cConf.set(Constants.AppFabric.TEMP_DIR, System.getProperty("java.io.tmpdir"));
-    cConf.setBoolean(Constants.Dangerous.UNRECOVERABLE_RESET, true);
 
     Injector injector = Guice.createInjector(
       new ConfigModule(cConf),
@@ -74,11 +65,6 @@ public class InMemoryStreamCoordinatorClientTest extends StreamCoordinatorTestBa
               })
     );
 
-    txManager = injector.getInstance(TransactionManager.class);
-    txManager.startAndWait();
-    datasetService = injector.getInstance(DatasetService.class);
-    datasetService.startAndWait();
-
     setupNamespaces(injector.getInstance(NamespacedLocationFactory.class));
     streamAdmin = injector.getInstance(StreamAdmin.class);
     coordinatorClient = injector.getInstance(StreamCoordinatorClient.class);
@@ -87,8 +73,6 @@ public class InMemoryStreamCoordinatorClientTest extends StreamCoordinatorTestBa
 
   @AfterClass
   public static void finish() {
-    datasetService.stopAndWait();
-    txManager.stopAndWait();
     coordinatorClient.stopAndWait();
   }
